WIP. Rip up topology, because clip algorithm was bad.
First problem was: pointer math was overkill. Array indices
seem more appropriate for us.

Second problem was: it's Really Hard to clip a topology, because
Sutherland-Hodgman can create invalid polygons, breaking the
assumptions we depend on. Instead, let's clip geometries.

Clipping topology instead of geometry stood to be twice as fast,
so it's a shame to drop that feature. Maybe we can revive it as
an optimization in the future, for Regions that are entirely
within the clip mask.

This is WIP because the actual clip algorithm hasn't been
implemented. There are tested support functions, though.